feat: Add SSL Policiy spec to the GCP LB Frontend (#15)
* feat: Add sslPolicy to gcp-load-balancer

This commit adds the `sslPolicy` field to the `gcp-load-balancer` component in the `Kptfile` and `frontendconfig.yaml` files. The value of `sslPolicy` is set to `gcp-load-balancer-ssl-policy` in both files. This change is made to configure the SSL policy for the load balancer.

* docs: Add Kpt documentation and configuration files

This commit adds the necessary documentation and configuration files for managing Kpt packages and blueprints. The `index.md` file provides an overview of Kpt packages and their usage. The `mkdocs.yml` file configures the site name and navigation for the documentation. These changes enable users to easily understand and utilize the Kpt packages and blueprints provided in this repository.

* chore: Update Kpt documentation and configuration files